fix: delete user_idindex on projects because its ceated through t.references

+++ db/schema.rb
@@ -18,7 +18,7 @@ ActiveRecord::Schema.define(version: 20180415011413) do
   create_table "projects", force: :cascade do |t|
     t.string "name", null: false
     t.text "description"
-    t.integer "user_id", null: false
+    t.bigint "user_id", null: false
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
     t.index ["user_id"], name: "index_projects_on_user_id"
